removed bug in update form & added django crispy forms
Solved a bug in the charity update form, which was causing the old data not to be populated on the form for updating, it was showing empty input fields. Solved this by referencing the profile update form of charities app. This solution will have to be replicated in the volunteers app profile update form. Also added django-crispy-forms to the project so that the forms are better styled. Added the crispy tags to the login and signup pages so far